Add basic auth views and templates
At least to start we will use the built-in auth User model, but
I want to write all my own templates and at least some of the
views and forms. In this commit I have started a 'core' app for
core site functionality, including user account handling. In there
I have written login, logout, home, index, and signup views and
templates, as well as some custom forms and the required changes
to urls.py.

In addition I have made a handful of changes to settings.py.