Handle redirection on signup and login
Rearrangement of core/forms.py, and add 'next' field to SignupForm.

Small corrections to login template to improve logic on checking
for a 'next' parameter in the form and change form target to the
default.

Add 'next' field to signup template.

Improve/refactor logic in core.views to handle redirects on signup
and follow DRY principles to put overlapping code of signup and
login into same function.