Set up password change and reset flows
Based on the builtin auth system password change and reset.

Password change is just about the same but with a custom template.

Password reset takes username in request instead of email. Token
handling is just about the same as in the builtin version, with
some minor differences, mainly in error handling.

The urls are slightly different from the builtin auth, and are
hopefully clearer. There are no confirmation urls after the
password reset request or successful change.

Customized templates all around, including the reset link email.

Some site settings were changed, with a couple new parameters
relating to the site domain/url/name, a manually set password
reset timeout period, and an email backend that prints to the
console (stdout) for development.