Create gearlists app, rename core -> accounts
Create the gearlists app, including models and basic forms. This
includes the GearItem, GearList, and GearListMembership models,
plus the GearItemManager class and the GearUserProfile model. The
manager class is there to handle batch checkout and checkin as
well as friendly unit display and perform other QuerySet tasks on
GearItems. The profile is to store user preferences and other info
in a one-to-one object for users.

signals.py - listen for User creation, create GearUserProfile

Objects are mostly nicely linked into the admin.

Update requirements.txt.

Rename `core` to `accounts`.

Reset db and create new migrations.